requirements.txt
absl-py==1.1.0
altair==4.2.0
astunparse==1.6.3
attrs==22.1.0
blinker==1.5
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
cycler==0.11.0
decorator==5.1.1
entrypoints==0.4
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.19.3
importlib-metadata==4.12.0
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
networkx==2.8.4
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
protobuf==3.19.4
pyarrow==10.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
pydicom==2.3.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
requests==2.28.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.8
scikit-image==0.19.3
scipy==1.8.1
semver==2.13.0
six==1.16.0
smmap==5.0.0
streamlit==1.15.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
tifffile==2022.5.4
toml==0.10.2
toolz==0.12.0
tornado==6.2
typing_extensions==4.2.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.9
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
